bot/requirements.txt
discord.py==2.3.2
python-dotenv
aiosqlite==0.19.0
wavelink==3.4.1
